requirements.txt
streamlit
pandas
numpy
altair
matplotlib.pyplot